Add files via upload
Setup & Configure
Create a package.json File:. In the terminal.
Run npm init -y to generate a package.json file. This file will manage your project’s dependencies.
Install Dependencies:
If you’re starting a new project, use npm install to install the necessary dependencies specified in your package.json
If the first byte is not 0 or 1, it returns false. Otherwise, it returns true, indicating that the signature is valid.
Committing main files. Including SmartContract file that includes function isValidSignature . it takes a hash and a signature as input, and returns a Boolean indicating whether the signature is valid or not.
The function first checks if the signature is not empty. If it is, it returns false.
Then, it gets the first byte of the signature using uint8(signature[0]).